requirements.txt
aiohttp==3.8.1
python-dotenv==0.21.0